feat(reporter): be able to use reporters in validations module
