+++ lib/reporters/ci-reporter.js
@@ -53,6 +53,7 @@ module.exports = {
     reportSucceededSequence,
     reportSucceededProcess,
     reportSuccess,
+    formatAsElegantPath,
 };
 
 /**
@@ -153,4 +154,18 @@ function reportSucceededSequence(message) {
 function reportSucceededProcess(message) {
     console.log('');
     console.log(message);
+}
+
+/**
+ * format input path
+ * @param {string} path
+ * @param {string} sep - path separator
+ * @returns {string} - returns input path as dir1 -> dir2 -> ... -> dirN -> file
+ */
+function formatAsElegantPath(path, sep) {
+    const packages = path.split(sep);
+
+    // prettier-ignore
+    const result = packages.map(item => item.trim()).join(' -> ');
+    return result;
 }
