+++ lib/reporters/current.js
@@ -40,4 +40,5 @@ module.exports = {
  * @property {function(string): function(boolean):void} reportRunningTask - report a task that is executing and may take some time,
  * @property {function(string): void} reportStep - report step message
  * @property {function(string): void} reportSuccess - report success message
+ * @property {function(string, string): string} formatAsElegantPath - report a successful execution of a sequence
  */
