+++ lib/reporters/elegant-status-reporter.js
@@ -54,6 +54,7 @@ module.exports = {
     reportSucceededSequence,
     reportSucceededProcess,
     reportSuccess,
+    formatAsElegantPath,
 };
 
 /**
@@ -153,4 +154,21 @@ function reportSucceededProcess(_message) {
     const emoji = require('node-emoji').emoji;
     console.log('');
     console.log(emoji.tada, emoji.tada, emoji.tada);
+}
+
+/**
+ * format input path
+ * @param {string} path
+ * @param {string} sep - path separator
+ */
+function formatAsElegantPath(path, sep) {
+    const chalk = require('chalk');
+    const packages = path.split(sep);
+    const lastIndex = packages.length - 1;
+
+    // prettier-ignore
+    const result = packages.map(item => item.trim()).map((item, index) => {
+        return index === lastIndex ? chalk.red.bold(item) : item;
+    }).join(' -> ');
+    return result;
 }
